+++ backend/src/main/java/org/rdfarchitect/services/validation/rule/ProfileHeaderValidationRule.java
@@ -19,7 +19,7 @@ package org.rdfarchitect.services.validation.rule;
 
 import org.apache.jena.graph.Graph;
 import org.apache.jena.graph.Node;
-import org.apache.jena.graph.NodeFactory;
+import org.apache.jena.graph.Triple;
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.vocabulary.RDF;
 import org.apache.jena.vocabulary.RDFS;
@@ -31,6 +31,8 @@ import org.rdfarchitect.api.dto.validation.SchemaValidationIssueDTO;
 import org.rdfarchitect.api.dto.validation.SchemaValidationIssueDTO.Severity;
 import org.rdfarchitect.models.cim.ontology.KnownOntologyFields;
 import org.rdfarchitect.models.cim.ontology.OntologyFactory;
+import org.rdfarchitect.models.cim.rdf.resources.CIMS;
+import org.rdfarchitect.models.cim.rdf.resources.CIMStereotypes;
 import org.springframework.stereotype.Component;
 
 import java.util.List;
@@ -42,43 +44,37 @@ public class ProfileHeaderValidationRule implements ValidationRule {
 
     private static final Set<OntologyField> REQUIRED_ONTOLOGY_FIELDS =
             Set.of(KnownOntologyFields.OWL_VERSION_IRI, KnownOntologyFields.DCAT_KEYWORD);
-
-    private static final String NS_CIMS = "http://iec.ch/TC57/1999/rdf-schema-extensions-19990926#";
-    private static final String CLASS_CLASS_CATEGORY = "ClassCategory";
     private static final String PACKAGE_FILE_HEADER_PROFILE = "#Package_FileHeaderProfile";
-
-    /** Stereotype value used by ENTSO-E for the CIM16 profile version (header) class. */
-    private static final String STEREOTYPE_ENTSOE = "Entsoe";
 
     /**
      * Local name suffix of the ENTSO-E profile version (header) class, e.g. DiagramLayoutVersion.
      */
     private static final String VERSION_SUFFIX = "Version";
 
-    private static final Node TYPE_CLASS_CATEGORY =
-            NodeFactory.createURI(NS_CIMS + CLASS_CLASS_CATEGORY);
-    private static final Node CIMS_STEREOTYPE = NodeFactory.createURI(NS_CIMS + "stereotype");
-
     @Override
     public void validate(
             Model model, List<SchemaValidationIssueDTO> issues, CGMESVersion cgmesVersion) {
         if (cgmesVersion == CGMESVersion.V3_0) {
-            var ontologyDTO = OntologyFactory.createOntologyDTO(model);
-
-            if (ontologyDoesNotExists(issues, ontologyDTO)) {
-                return;
-            }
-
-            var presentIris =
-                    ontologyDTO.getEntries().stream()
-                            .map(OntologyEntry::getIri)
-                            .collect(Collectors.toSet());
-
-            validateRequiredFields(issues, presentIris, ontologyDTO);
-            validateOptionalFields(issues, presentIris, ontologyDTO);
+            validate3_0ProfileHeader(model, issues);
         } else if (cgmesVersion == CGMESVersion.V2_4_15) {
             validate2_4_15ProfileHeader(model, issues);
         }
+    }
+
+    private void validate3_0ProfileHeader(Model model, List<SchemaValidationIssueDTO> issues) {
+        var ontologyDTO = OntologyFactory.createOntologyDTO(model);
+
+        if (ontologyDoesNotExists(issues, ontologyDTO)) {
+            return;
+        }
+
+        var presentIris =
+                ontologyDTO.getEntries().stream()
+                        .map(OntologyEntry::getIri)
+                        .collect(Collectors.toSet());
+
+        validateRequiredFields(issues, presentIris, ontologyDTO);
+        validateOptionalFields(issues, presentIris, ontologyDTO);
     }
 
     private static void validate2_4_15ProfileHeader(
@@ -90,7 +86,7 @@ public class ProfileHeaderValidationRule implements ValidationRule {
                             .severity(Severity.WARNING)
                             .message(
                                     "Profile header is missing: no Package_FileHeaderProfile"
-                                            + " and no ENTSO-E profile version header found.")
+                                            + " or ENTSO-E profile version header found.")
                             .build());
         }
     }
@@ -153,7 +149,7 @@ public class ProfileHeaderValidationRule implements ValidationRule {
     }
 
     public static boolean isCim16HeaderProfile(Graph graph) {
-        return graph.stream(Node.ANY, RDF.type.asNode(), TYPE_CLASS_CATEGORY)
+        return graph.stream(Node.ANY, RDF.type.asNode(), CIMS.classCategory.asNode())
                 .anyMatch(
                         t ->
                                 t.getSubject().isURI()
@@ -173,26 +169,13 @@ public class ProfileHeaderValidationRule implements ValidationRule {
      */
     public static boolean hasCim16VersionHeader(Graph graph) {
         return graph.stream(Node.ANY, RDF.type.asNode(), RDFS.Class.asNode())
-                .map(org.apache.jena.graph.Triple::getSubject)
+                .map(Triple::getSubject)
                 .filter(Node::isURI)
-                .filter(subject -> localName(subject.getURI()).endsWith(VERSION_SUFFIX))
+                .filter(subject -> subject.getURI().endsWith(VERSION_SUFFIX))
                 .anyMatch(subject -> hasEntsoeStereotype(graph, subject));
     }
 
     private static boolean hasEntsoeStereotype(Graph graph, Node subject) {
-        return graph.stream(subject, CIMS_STEREOTYPE, Node.ANY)
-                .map(org.apache.jena.graph.Triple::getObject)
-                .anyMatch(
-                        object ->
-                                object.isLiteral()
-                                        && STEREOTYPE_ENTSOE.equals(
-                                                object.getLiteralLexicalForm()));
-    }
-
-    private static String localName(String uri) {
-        int hash = uri.lastIndexOf('#');
-        int slash = uri.lastIndexOf('/');
-        int idx = Math.max(hash, slash);
-        return idx >= 0 && idx < uri.length() - 1 ? uri.substring(idx + 1) : uri;
+        return graph.contains(subject, CIMS.stereotype.asNode(), CIMStereotypes.entsoe.asNode());
     }
 }
